requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-qdrant
langchain-text-splitters
qdrant-client
huggingface-hub
sentence-transformers
pypdf
PyPDF2
requests
python-multipart
streamlit
pydantic-settings
uuid-utils
SpeechRecognitionf a s t e m b e d  
 